compile success, run incorrect

+++ lua-5.2.0-2009/kopilua/lstring.h.cs
@@ -6,7 +6,7 @@ namespace KopiLua
 	{
 		public static int sizestring(TString s) {return ((int)s.len + 1) * GetUnmanagedSize(typeof(char)); }
 
-		public static int sizeudata(Udata u) { return (int)u.len; }
+		public static uint sizeudata(Udata u) { return u.len; }
 
 		public static TString luaS_new(lua_State L, CharPtr s) { return luaS_newlstr(L, s, (uint)strlen(s)); }
 		public static TString luaS_newliteral(lua_State L, CharPtr s) { return luaS_newlstr(L, s, (uint)strlen(s)); }
